requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
python-dotenv